Get rid of ExceptT, compat w/ new http-conduit & monad-metrics
http-conduit now requires MonadUnliftIO:
https://github.com/snoyberg/http-client/commit/dfbcb6c28a3216d0a69adfa9ccc8bdf62aff974d

It's possible to work around that by running the http client
body handlers in a fresh ReaderT. But that's an extra limitation.

monad-metrics now requires MonadMask:
https://github.com/parsonsmatt/monad-metrics/commit/17546b92b4e7e94279b81afe76fd6daa5f3ff0f8

It's *not* possible to work around if you want to measure whole
handlers.

So, ExceptT is slower anyway, and an antipattern:
https://www.parsonsmatt.org/2017/06/21/exceptional_servant_handling.html

Now we're free of it! And exporting UnliftIO by default.